Implement AzureAD OAuth2 backend
This adds an additional OAuth2 backend named `azure` and the
corresponding documentation. To make the backend work, this also
modifies the JWK validation code to allow a missing `alg` header,
which is optional according to the spec and not present in Microsoft's
JWKs.

I hereby consent to the Individual CLA provided in
assets/cla/individual_cla.md